+++ scripts/__tests__/size-base-cache.test.ts
@@ -86,13 +86,39 @@ test('removal cannot run at all while another run holds the takeover mutex', () 
   release();
 });
 
-test('a takeover mutex abandoned by a killed process is reclaimed by age, not wedged forever', () => {
+test('a delayed takeover holder is never displaced, however old its mutex looks', () => {
+  // The interleaving that age-based reclamation created: holder A is merely slow — paused, or
+  // SIGSTOPed past any threshold — while still inside the section. Reclaiming its mutex would put
+  // B inside too, and then A's release could remove B's mutex and either could unlink the claim
+  // the other just created. A mutex is therefore never taken from its holder, at any age.
   fs.symlinkSync(ABANDONED, claimPath(entry));
-  fs.mkdirSync(takeoverPath(entry));
-  const longAgo = new Date(Date.now() - 60_000);
-  fs.utimesSync(takeoverPath(entry), longAgo, longAgo);
-  const release = acquireBaseClaim(entry, 'abc123456');
-  assert.equal(readClaimIdentity(claimPath(entry)), CLAIM_IDENTITY);
+  fs.symlinkSync(`${process.pid}:delayed-holder`, takeoverPath(entry));
+  const longAgo = new Date(Date.now() - 24 * 60 * 60 * 1000);
+  fs.lutimesSync(takeoverPath(entry), longAgo, longAgo);
+
+  assert.equal(removeAbandonedClaim(entry, ABANDONED), 'busy');
+  assert.equal(
+    fs.readlinkSync(takeoverPath(entry)),
+    `${process.pid}:delayed-holder`,
+    "the holder's mutex is intact",
+  );
+  assert.equal(readClaimIdentity(claimPath(entry)), ABANDONED, 'and it removed nothing');
+  assert.throws(() => acquireBaseClaim(entry, 'abc123456'), /taking over the abandoned claim/);
+  assert.equal(fs.readlinkSync(takeoverPath(entry)), `${process.pid}:delayed-holder`);
+});
+
+test('a leaked mutex wedges only its own entry, and says how to clear it', () => {
+  // The price of never reclaiming: one entry needs a human. The message has to name the path.
+  fs.symlinkSync(ABANDONED, claimPath(entry));
+  fs.symlinkSync(`${NEVER_A_PID}:leaked`, takeoverPath(entry));
+  assert.throws(
+    () => acquireBaseClaim(entry, 'abc123456'),
+    new RegExp(`remove ${takeoverPath(entry).replace(/[.*+?^${}()|[\]\\]/g, '\\$&')}`),
+  );
+  // A different entry is unaffected: the mutex is per entry, so nothing else is wedged.
+  const other = path.join(path.dirname(entry), 'def987654321');
+  const release = acquireBaseClaim(other, 'def987654');
+  assert.equal(readClaimIdentity(claimPath(other)), CLAIM_IDENTITY);
   release();
 });
